api/requirements.txt
# Grandmaster Guard - Python Backend

fastapi>=0.104.0
uvicorn>=0.24.0
python-chess>=1.999
httpx>=0.25.0
pymongo>=4.6.0
motor>=3.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
chess.com>=3.13.0
openai>=1.0.0
google-generativeai>=0.8.0
zstandard>=0.22.0
requests>=2.31.0

# LangGraph AI Coach
langgraph>=0.2.0
langchain-ollama>=0.2.0
langchain-core>=0.3.0

# Video Content Pipeline
edge-tts>=6.1.0
